refactor: extract ResourceClassLoader and ModelManifestManager from StorytellingResources
Decompose StorytellingResources.java from 611 lines to 361 lines (target: <500)
by extracting two focused helper classes:

- ResourceClassLoader (210 lines): Static utility for discovering and loading
  ModelResource classes from jar/directory trees. Contains getClassNamesFromResources,
  loadClassesFromResourceFiles, getAndLoadModelResourceClasses, and LoadResult value
  class that decouples class loading from instance state.

- ModelManifestManager (170 lines): Stateful manager owning lazy-loaded caches of
  ModelManifest instances. Contains getDynamicModelFiles, findAndLoadUserGalleryResources,
  findAndLoadInternalResources, findNewUserGalleryResources, and manifest lookup methods.

StorytellingResources remains the public facade with unchanged API signatures.
Key changes:
- DIR_FILE_FILTER anonymous class replaced with File::isDirectory lambda
- Dead code removed (commented-out getPathFromProperties, DEBUG static block)
- //TEST dead code in loadClassesFromResourceFiles removed
- findAndLoadInstalledAliceResourcesIfNecessary uses LoadResult to unpack classes
  and classloaders from the now-static ResourceClassLoader methods

All 74 decomposition tests pass. NebulousStorytellingResources compiles unchanged.

Closes #648
